Refactor report model to use AdminStatus enum and update related classes for improved type safety

+++ homemate-backend/src/main/java/com/homemate/util/PaginatedResponse.java
@@ -13,9 +13,9 @@ import lombok.Builder;
 @Builder
 public class PaginatedResponse<T> {
     private List<T> data;
-    private int currentPage;
-    private int totalPages;
+    private long currentPage;
+    private long totalPages;
     private long totalElements;
-    private int pageSize;
+    private long pageSize;
 }
